Update library dependencies #171
Oshi changes group from dblock to oshi to get to most recent 3.12.0

purejavacomm also changes group to org.jetbrains.pty4j and should have less
dependencies anyway. We need the jcenterRepo resolver because of this change.
It is unknown whether the newer purejavacomm is fully compatible to the old
one.