Correct and simplify DefaultAhbLite3Slave
There is no need to have 4 states here, actually the AHB-Lite 
specification requires an error response to last for exactly two clock 
cycles, the first with HREADYOUT='0' and HRESP='1', and the second with 
HREADYOUT='1' and HRESP='1'. The state machine doesn’t work correctly if 
it is in ERROR state with a new request arriving, which is only 
considered in IDLE state. Also, we should be prepared for new requests 
at any time, so we should decode io.HREADY input as well and restart the 
sequence irrespective of our internal state.